# docs/architecture/openapi.phase1.yaml
openapi: 3.0.3
info:
  title: PawMate Phase 1 API
  version: 1.0.0
  description: Phase 1 contract for auth, pets, vets, reviews, health records, reminders, and notifications.
servers:
  - url: http://localhost:3000
  - url: https://api.pawmate.example
...
security:
  - bearerAuth: []
paths:
...
  /vets/nearby:
    get:
      tags: [Vets]
      summary: Search nearby clinics
      parameters:
        - in: query
          name: lat
          schema: { type: number, format: float }
        - in: query
          name: lng
          schema: { type: number, format: float }
        - in: query
          name: radiusKm
          schema: { type: integer, default: 5, minimum: 1, maximum: 50 }
        - in: query
          name: openNow
          schema: { type: boolean }
        - in: query
          name: emergency24h
          schema: { type: boolean }
        - in: query
          name: minRating
          schema: { type: number, format: float, minimum: 0, maximum: 5 }
      responses:
        '200':
          description: Nearby clinics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetList'
  /vets/search:
    get:
      tags: [Vets]
      summary: Search clinics by text and filters
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: service
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetList'
  /vets/{vetId}:
    parameters:
      - $ref: '#/components/parameters/VetId'
    get:
      tags: [Vets]
      summary: Get clinic detail
      responses:
        '200':
          description: Clinic detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetItem'
  /vets/{vetId}/reviews:
    parameters:
      - $ref: '#/components/parameters/VetId'
    get:
      tags: [Reviews]
      summary: List reviews for a clinic
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
    post:
      tags: [Reviews]
      summary: Create a clinic review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpsert'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    VetId:
      in: path
      name: vetId
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Meta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    Vet:
      type: object
      required: [id, name, address, distanceKm]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        address: { type: string }
        distanceKm: { type: number, format: float }
        rating: { type: number, format: float, nullable: true }
        reviewCount: { type: integer }
        is24h: { type: boolean }
        openNow: { type: boolean }
        services:
          type: array
          items: { type: string }
    VetItem:
      type: object
      required: [data]
      properties:
        data:
          type: object
          allOf:
            - $ref: '#/components/schemas/Vet'
            - type: object
              properties:
                latitude: { type: number, format: float }
                longitude: { type: number, format: float }
                hours: { type: object, additionalProperties: true }
                reviewSummary: { type: object, additionalProperties: true }
    VetList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Vet' }
        meta: { $ref: '#/components/schemas/Meta' }
    Review:
      type: object
      required: [id, vetId, userId, rating, createdAt]
      properties:
        id: { type: string, format: uuid }
        vetId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        title: { type: string, nullable: true }
        body: { type: string, nullable: true }
        isAnonymous: { type: boolean }
        createdAt: { type: string, format: date-time }
    ReviewUpsert:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        title: { type: string, nullable: true }
        body: { type: string, nullable: true }
        isAnonymous: { type: boolean, default: false }
    ReviewItem:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Review' }
    ReviewList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Review' }
        meta: { $ref: '#/components/schemas/Meta' }
...